+++ java/src/main/kotlin/io/spine/validation/java/generate/option/MaxGenerator.kt
@@ -26,15 +26,34 @@
 
 package io.spine.validation.java.generate.option
 
+import io.spine.base.FieldPath
 import io.spine.protodata.ast.TypeName
+import io.spine.protodata.ast.name
+import io.spine.protodata.java.Expression
+import io.spine.protodata.java.StringLiteral
 import io.spine.server.query.Querying
 import io.spine.server.query.select
+import io.spine.validation.MAX
 import io.spine.validation.bound.MaxField
+import io.spine.validation.bound.NumericBound
+import io.spine.validation.bound.NumericBound.ValueCase.UINT32_VALUE
+import io.spine.validation.bound.NumericBound.ValueCase.UINT64_VALUE
+import io.spine.validation.java.expression.IntegerClass
+import io.spine.validation.java.expression.LongClass
+import io.spine.validation.java.expression.joinToString
+import io.spine.validation.java.expression.stringValueOf
 import io.spine.validation.java.generate.OptionCode
 import io.spine.validation.java.generate.OptionGenerator
+import io.spine.validation.java.violation.ErrorPlaceholder
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.java.violation.ErrorPlaceholder.MAX_OPERATOR
+import io.spine.validation.java.violation.ErrorPlaceholder.MAX_VALUE
+import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
 
 /**
- * The generator for `(max)` option.
+ * The generator for the `(max)` option.
  */
 internal class MaxGenerator(private val querying: Querying) : OptionGenerator {
 
@@ -49,5 +68,45 @@ internal class MaxGenerator(private val querying: Querying) : OptionGenerator {
     override fun codeFor(type: TypeName): List<OptionCode> =
         allMaxFields
             .filter { it.id.type == type }
-            .map { MaxFieldGenerator(it).generate() }
+            .map { GenerateMax(it).code() }
+}
+
+/**
+ * Generates code for a single application of the `(max)` option
+ * represented by the [view].
+ */
+private class GenerateMax(private val view: MaxField) : BoundedFieldGenerator(view, MAX) {
+
+    private val bound = view.bound
+    private val isExclusive = bound.exclusive
+
+    override val boundPrimitive: NumericBound.ValueCase = bound.valueCase
+
+    /**
+     * Returns a boolean expression that checks if the given [value] exceeds
+     * the maximum [bound].
+     */
+    @Suppress("MaxLineLength") // Easier to read.
+    override fun isOutOfBounds(value: Expression<Number>): Expression<Boolean> {
+        val literal = bound.asLiteral()
+        val operator = if (isExclusive) ">=" else ">"
+        return when (boundPrimitive) {
+            UINT32_VALUE -> Expression("$IntegerClass.compareUnsigned($value, $literal) $operator 0")
+            UINT64_VALUE -> Expression("$LongClass.compareUnsigned($value, $literal) $operator 0")
+            else -> Expression("$value $operator $literal")
+        }
+    }
+
+    override fun supportedPlaceholders(
+        fieldPath: Expression<FieldPath>,
+        typeName: Expression<String>,
+        fieldValue: Expression<*>,
+    ): Map<ErrorPlaceholder, Expression<String>> = mapOf(
+        FIELD_PATH to fieldPath.joinToString(),
+        FIELD_VALUE to fieldType.stringValueOf(fieldValue),
+        FIELD_TYPE to StringLiteral(fieldType.name),
+        PARENT_TYPE to typeName,
+        MAX_VALUE to StringLiteral(view.max),
+        MAX_OPERATOR to StringLiteral(if (isExclusive) "<" else "<=")
+    )
 }
